Pull PascalCaseToSnakeCase into the util package.
